Dev/205 cleanup (#264)
* cleanup from refactor

* readme link around

* readme fmt
