+++ .github/CODEOWNERS
@@ -1,1 +1,1 @@
-* @abdelhamidbakhta @LucasLvy
+* @drspacemn @danilowhk
