+++ .github/workflows/stale.yml
@@ -1,5 +1,5 @@
 ---
-name: Stale
+name: stale
 
 on:
   schedule:
