+++ crates/beerus-core/Cargo.toml
@@ -23,13 +23,3 @@ tokio = { version = "1.21.2", features = ["macros"] }
 [dev-dependencies]
 httpmock = "0.6.7"
 temp-env = "0.3.1"
-
-[dev-dependencies.cargo-husky]
-version = "1.5.0"
-default-features = false
-features = [
-    "precommit-hook",
-    "run-for-all",
-    "run-cargo-clippy",
-    "run-cargo-fmt",
-]
