+++ crates/beerus-rpc/README.md
@@ -1,0 +1,4 @@
+# Beerus RPC Usage
+
+Beerus RPC provides a variety of endpoints to interact with via the
+StarkNet and Ethereum Networks
