--- src/lib.rs
@@ -1,3 +1,0 @@
-pub mod beerus_core {
-    pub use beerus_core::{config::Config, lightclient::beerus::BeerusLightClient};
-}
